+++ crates/cli/src/commands/collections.rs
@@ -1,7 +1,15 @@
-use crate::{GlobalArgs, Subcommand, util::print_table};
+use crate::{
+    GlobalArgs, Subcommand, completions::complete_collection_specifier,
+    util::print_table,
+};
 use clap::Parser;
-use slumber_core::database::Database;
-use std::{path::PathBuf, process::ExitCode};
+use slumber_core::database::{CollectionId, Database};
+use std::{
+    fmt::{self, Display},
+    path::PathBuf,
+    process::ExitCode,
+    str::FromStr,
+};
 
 /// View and modify request collection metadata
 #[derive(Clone, Debug, Parser)]
@@ -20,10 +28,12 @@ enum CollectionsSubcommand {
     /// The data from the source collection will be merged into the target
     /// collection, then all traces of the source collection will be deleted!
     Migrate {
-        /// The path the collection to migrate *from*
-        from: PathBuf,
-        /// The path the collection to migrate *into*
-        to: PathBuf,
+        /// Path or ID the collection to migrate *from*
+        #[clap(add = complete_collection_specifier())]
+        from: CollectionSpecifier,
+        /// Path or ID the collection to migrate *into*
+        #[clap(add = complete_collection_specifier())]
+        to: CollectionSpecifier,
     },
 }
 
@@ -45,10 +55,47 @@ impl Subcommand for CollectionsCommand {
                 print_table(["Path", "Name"], &rows);
             }
             CollectionsSubcommand::Migrate { from, to } => {
-                database.merge_collections(&from, &to)?;
-                println!("Migrated {} into {}", from.display(), to.display());
+                let from_id = from.to_id(&database)?;
+                let to_id = to.to_id(&database)?;
+                database.merge_collections(from_id, to_id)?;
+                println!("Migrated {from} into {to}");
             }
         }
         Ok(ExitCode::SUCCESS)
+    }
+}
+
+/// Specify a collection by file path or ID
+#[derive(Clone, Debug)]
+enum CollectionSpecifier {
+    Id(CollectionId),
+    Path(PathBuf),
+}
+
+impl CollectionSpecifier {
+    fn to_id(&self, database: &Database) -> anyhow::Result<CollectionId> {
+        match self {
+            Self::Id(id) => Ok(*id),
+            Self::Path(path) => database.get_collection_id(path),
+        }
+    }
+}
+
+impl FromStr for CollectionSpecifier {
+    type Err = anyhow::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Ok(s.parse::<CollectionId>()
+            .map(Self::Id)
+            .unwrap_or_else(|_| Self::Path(PathBuf::from(s))))
+    }
+}
+
+impl Display for CollectionSpecifier {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Id(id) => write!(f, "{id}"),
+            Self::Path(path) => write!(f, "{}", path.display()),
+        }
     }
 }