+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased] - ReleaseDate
 
+### Changed
+
+- `slumber collections migrate` now accepts collection IDs in addition to paths
+
 ### Fixed
 
 - Shell completion for the global `--file`/`-f` flag only suggests `.yml`/`.yaml` files